Use `webpack` to build
`jison` requires `fs` module and applications using `sparql-connect` in the browser needed to add an option in their `webpack.config.js` to avoid an error while resolving `fs` module. So we switched to `webpack` here (instead of regular `babel` script), with this `node.fs` option, so other applications do not need to care about it.